Remove `aws_sdk_dynamodb` from common.
Only the "cloud" version of the provisioner needs to delete tables if the dynamodb resource is deleted.
